Teacher controller and method implemented, with thesis application submission. In addition, fix on all "id" parameters have been done.

+++ src/main/java/com/project/graduation/services/ThesisService.java
@@ -5,5 +5,5 @@ import org.springframework.web.multipart.MultipartFile;
 
 public interface ThesisService {
 
-    void saveThesis(ThesisDto thesisDto, int studentId, MultipartFile file);
+    void saveThesis(ThesisDto thesisDto, long studentId, MultipartFile file);
 }
